agents/voice_agent/requirements.txt
fastapi
uvicorn
openai-whisper
pyttsx3
python-multipart
pydantic